Revert "Move caas operator upgrader from operator manifold to model manifold"
This reverts commit 22165c07826bf970712ad008976c953327915af4.
